MDAnalysis InterRDF requires box dimensions
Added in storing the box dimensions into the COM extraction. Will
rename the file as it is not specifically just COMS anymore.

Also have a basic framework for the RDF calculation with MDAnalysis
now so I just need to put it all into a function.

Will need to investigate bins, and the range as well as other norms.