RDF calculator function producd plus minor changes
The RDF function chains together existing functions to calculate the
RDF. Need to come up with a method for plotting as there will be
many more RDFs.

Also changed the utility functions. Names are more representative of
the function purposes and carried edited so not putting `.npz` in
the output file will not throw errors in the RDF calculator function.