Polymer-polymer radial distribution function
Found that I don't actually need to extract the positions of the
reference particle for each frame. This saves me turning this
position array into a dummy `Universe`. Will investigate if I
still need the dummy Universe for the API centres of mass as this
will reduce a lot of code.