Minimal plotting function for plotting MSD curves
This will be used for plotting MSD curves at 300K in hopes of seeing
if the diffusion coefficient has converged at this temperature.

Again need to come up with a better mechanism to let the user know
that they are dealing with trajectories at 300 K - worry about that
once convergence is obtained.