MDA agrees with both normalisation schemes
Because box fluctuations are so small, both normalisation schemes agree,
hiding what MDAnalysis is doing. Read up and maybe test with a very
small simulation. Large pressure, just API for a short number of steps.

After reading the source code it looks like MDA is using the
averge density to normalise the RDF :(.