Some big changes to the RDF calculation
Only skip frames implemented into the rdf. This means no more
calculating centres of mass after the the equilbration time.

After the supervisor meeting on 06.10.25 we decided that the
manually coded RDFs look good. We discussed small fluctuations caused
the instantaneous and average normalisation to be very close.

We have also decicded to stop using MDAnalysis as it is quite
limited in what we can do. Use it only to extract atoms.

Next steps:
1. Remove the dead MDAnalysis cod
2. Tidy the remaining code with typeguarding
3. While more cooling ramps one, obtain the wet convergence and
the running coordination number.